Project 1
Classes in C++

This semester we will have some fun setting up our very own role playing game (inspired by the Divinity series). We hope you'll enjoy this adventure as much as we did setting it up for you!

This is a baseline project whose objective is to get you acquainted with the platforms we will use in this course and to refresh your knowledge of basic OOP. You will implement the Character class. In order to successfully complete this project, we strongly recommend that you look back to your CSCI 135 coursework as a reference.