Project 2
Inheritance in C++

You will now define different roles for your character. You will implement 4 new classes that derive from the Character class but also have additional attributes and methods specific to their type: 

Mages are scholars of magic and spells, and are able to manipulate the world around them with elemental magic. Each Mage has an affinity to a specific school of magic, a choice of weapon, and an optional choice of summoning an incarnate.

Scoundrels live by their wits, take chances when they must, and land on their feet against all odds. Every Scoundrel has a set of daggers made from a certain material (e.g. wood, bronze, iron, steel, mithril, adamant, and rune) faction that they belong to: Cutpurse, Nightblades, or Silvertongue (e.g, thieves, assassins, and swindlers). Scoundrels may or may not have a disguise.

Rangers specialize in hunting the monsters that threaten the edges of civilization. Each Ranger is equipped with a crossbow that fires from a limited source of ammunition. Arrows can come with different effects, so a ranger's ammunition source includes a range of different arrows. A Ranger may or may not be able to recruit an animal companion. Each Ranger also has a list of affinities and advantages, which increases the damage dealt with an arrow of the same element. 

Barbarians are the archetypal warriors who use brute strength and raw fury to excel in combat. A Barbarian can wield either a two-handed weapon, or a one-handed weapon and a shield. They may or may not be enraged, which allows them to deal double damage.